refactor: rename from FacilityPanel to FacilityDetailPanel
Add FacilityDetailRow
