+++ UI/src/features/owner/components/Wizard/Step4/FacilityDetailRow.tsx
@@ -1,13 +1,11 @@
-import { Box, Typography, Paper, Chip, Collapse, TextField, Button } from "@mui/material";
+import { Box, Typography, TextField, Button, Collapse } from "@mui/material";
 import ExpandMoreIcon from "@mui/icons-material/ExpandMore";
 import DeleteOutlineIcon from "@mui/icons-material/DeleteOutline";
 import HomeIcon from "@mui/icons-material/Home";
 import WifiIcon from "@mui/icons-material/Wifi";
 import PoolIcon from "@mui/icons-material/Pool";
 import LocalParkingIcon from "@mui/icons-material/LocalParking";
-import type { WizardForm, FacilityConfig } from "../../../types/owner.types";
-
-// ─── Icon mapper (shared, mirrors StepFacilityBox) ────────────────────────────
+import type { FacilityConfig } from "../../../types/owner.types";
 
 const getFacilityIcon = (name: string) => {
 	const key = name.toLowerCase();
@@ -17,8 +15,6 @@ const getFacilityIcon = (name: string) => {
 	return HomeIcon;
 };
 
-// ─── Single facility row ──────────────────────────────────────────────────────
-
 interface FacilityRowProps {
 	facility: FacilityConfig;
 	isExpanded: boolean;
@@ -27,7 +23,7 @@ interface FacilityRowProps {
 	onRemove: () => void;
 }
 
-function FacilityRow({ facility, isExpanded, onToggle, onUpdate, onRemove }: FacilityRowProps) {
+export const FacilityDetailRow: React.FC<FacilityRowProps> = ({ facility, isExpanded, onToggle, onUpdate, onRemove }) => {
 	const hasMeta = (facility.fee ?? 0) > 0 || !!facility.note;
 	const Icon = getFacilityIcon(facility.name);
 
@@ -150,91 +146,4 @@ function FacilityRow({ facility, isExpanded, onToggle, onUpdate, onRemove }: Fac
 			</Collapse>
 		</Box>
 	);
-}
-
-// ─── Panel ────────────────────────────────────────────────────────────────────
-
-interface Props {
-	form: WizardForm;
-	setForm: React.Dispatch<React.SetStateAction<WizardForm>>;
-	// expandedId is lifted here so StepFacilityBox can auto-expand on select
-	expandedId: string | null;
-	setExpandedId: React.Dispatch<React.SetStateAction<string | null>>;
-}
-
-export default function FacilityPanel({ form, setForm, expandedId, setExpandedId }: Props) {
-	const updateFacility = (id: string, patch: Partial<Pick<FacilityConfig, "fee" | "note">>) => {
-		setForm((prev) => ({
-			...prev,
-			facilities: prev.facilities.map((f) => (f.id === id ? { ...f, ...patch } : f)),
-		}));
-	};
-
-	const removeFacility = (id: string) => {
-		setForm((prev) => ({
-			...prev,
-			facilities: prev.facilities.filter((f) => f.id !== id),
-		}));
-		if (expandedId === id) setExpandedId(null);
-	};
-
-	return (
-		<Paper
-			elevation={0}
-			sx={{
-				width: 280,
-				flexShrink: 0,
-				borderRadius: 3,
-				border: "1px solid",
-				borderColor: "divider",
-				overflow: "hidden",
-				position: "sticky",
-				top: 24,
-			}}
-		>
-			{/* Header */}
-			<Box
-				sx={{
-					px: 2,
-					py: 1.5,
-					borderBottom: "1px solid",
-					borderColor: "divider",
-					display: "flex",
-					alignItems: "center",
-					justifyContent: "space-between",
-				}}
-			>
-				<Typography variant="subtitle2" fontWeight={700}>
-					Selected Facilities
-				</Typography>
-				{form.facilities.length > 0 && <Chip label={form.facilities.length} size="small" color="primary" />}
-			</Box>
-
-			{/* Empty state */}
-			{form.facilities.length === 0 && (
-				<Box px={2} py={4} textAlign="center">
-					<Typography variant="body2" color="text.disabled">
-						No facilities selected yet.
-					</Typography>
-					<Typography variant="caption" color="text.disabled">
-						Pick tiles on the left to add them here.
-					</Typography>
-				</Box>
-			)}
-
-			{/* Rows in selection order */}
-			<Box sx={{ display: "flex", flexDirection: "column", gap: 1, p: form.facilities.length > 0 ? 1.5 : 0 }}>
-				{form.facilities.map((f) => (
-					<FacilityRow
-						key={f.id}
-						facility={f}
-						isExpanded={expandedId === f.id}
-						onToggle={() => setExpandedId((prev) => (prev === f.id ? null : f.id))}
-						onUpdate={(patch) => updateFacility(f.id, patch)}
-						onRemove={() => removeFacility(f.id)}
-					/>
-				))}
-			</Box>
-		</Paper>
-	);
-}
+};
